backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
branca==0.8.1
category_encoders==2.7.0
certifi==2025.7.14
charset-normalizer==3.4.2
choreographer==1.0.9
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
dash==3.1.1
decorator==5.2.1
deprecation==2.1.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
fastjsonschema==2.21.1
Flask==3.1.1
folium==0.20.0
fonttools==4.58.5
gurobipy==12.0.2
h11==0.16.0
httptools==0.6.4
idna==3.10
imbalanced-learn==0.13.0
importlib_metadata==8.7.0
ipython==8.37.0
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.3.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kaleido==1.0.0
kiwisolver==1.4.8
lightgbm==4.6.0
llvmlite==0.44.0
logistro==1.1.0
MarkupSafe==3.0.2
matplotlib==3.7.5
matplotlib-inline==0.1.7
nbformat==5.10.4
nest-asyncio==1.6.0
numba==0.61.2
numpy==1.26.4
orjson==3.10.18
packaging==25.0
pandas==2.1.4
parso==0.8.4
patsy==1.0.1
pillow==11.3.0
pip==25.1.1
platformdirs==4.3.8
plotly==5.24.1
plotly-resampler==0.10.0
pmdarima==2.0.4
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pycaret==3.3.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyod==2.0.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==310
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
retrying==1.4.0
rpds-py==0.26.0
rtree==1.4.0
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
setuptools==80.9.0
shap==0.48.0
shapely==2.1.1
simplejson==3.20.1
six==1.17.0
sklearn-compat==0.1.3
sktime==0.26.0
slicer==0.0.8
sniffio==1.3.1
stack-data==0.6.3
starlette==0.47.1
statsmodels==0.14.5
tbats==1.1.3
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
traitlets==5.14.3
tsdownsample==0.1.4.1
typing_extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
xxhash==3.5.0
xyzservices==2025.4.0
yellowbrick==1.5
zipp==3.23.0
